requirements.txt
pyspark==3.3.0
minio==7.1.13
psycopg2-binary==2.9.5
pandas==1.5.3
numpy==1.23.5
pyyaml==6.0
pyarrow==10.0.1
delta-spark==2.3.0
loguru==0.6.0
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0 
